Add to AudioDevice API and implement it for PortaudioAudioDevice
Improve portaudio/wscript style

+++ src/mojo/core/typesystem/types.hpp
@@ -8,7 +8,7 @@
 namespace mojo {
 
 typedef int64_t     count_t;
-typedef uint8_t     channel_count_t;
+typedef uint32_t    channel_count_t;
 typedef uint32_t    samplerate_t;
 
 } // namespace mojo
